Some cleanup of CSDiag by splitting off the CalleeCandidateInfo class and declaring a few formerly static methods that both files now need in a new CSDiag.h.
