+++ lib/Sema/CMakeLists.txt
@@ -12,6 +12,7 @@ add_swift_library(swiftSema STATIC
   CSRanking.cpp
   CSSimplify.cpp
   CSSolver.cpp
+  CalleeCandidateInfo.cpp
   CodeSynthesis.cpp
   Constraint.cpp
   ConstraintGraph.cpp
